+++ TrabalhoFinal/03-Entidade/Venda.cs
@@ -14,7 +14,7 @@ namespace Core._03_Entidades
         public int PessoaId { get; set; }
         public int ProdutoId { get; set; }
         public int MetodoDePagamento { get; set; }
-        public double ValorFinal { get; set; }
+        public decimal ValorFinal { get; set; }
 
         public override string ToString()
         {
